Merged PR 18: Add the Azure Functions to update the Status of the device + Unit Tests
The most relevant Azure Function is `UpdateDeviceTwinStatusProperty`:
- it updates the reported property (Status) on the Device Twin.
- it is covered with Unit Tests.

The project uses StyleCop and Dependency Injection (to make unit testing possible).

Note: `RentDevice/ReturnDevie/RechargeDevice` were created for the scenario in which the properties of the device were stored in memory. In that case, direct methods were invoked - which were registered on the Console Application. However, I moved from that scenario to storing the properties on the Device Twin. Therefore, those are not covered with tests and were left only for learning purpose.

Related work items: #110